requirements.txt
fastapi
peft
pydantic
python-dotenv
qdrant-client
sentence-transformers
torch
transformers
uvicorn
sentencepiece
bitsandbytes